feat: terminados os testes para o módulo de histórico, todos os ttestes estão funcionando sem erros

+++ src/main/java/com/project/ecopraia/service/HistoricoService.java
@@ -54,6 +54,10 @@ public class HistoricoService {
     }
 
     public void excluir(Long id) {
+        if(!historicoRepository.existsById(id)) {
+            throw  new RuntimeException("Historico não encontrado");
+        }
+        
         historicoRepository.deleteById(id);
     }
 }
